[round-4] 과제 제출 - 서상원

+++ apps/commerce-api/src/main/java/com/loopers/interfaces/api/point/PointV1Dto.java
@@ -9,7 +9,7 @@ public class PointV1Dto {
     ) {
         public static PointResponse from(UserPointInfo userPointInfo) {
             return new PointResponse(
-                    userPointInfo.point()
+                    userPointInfo.point().intValue()
             );
         }
     }
